# src/admin-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/accounts": {
      "post": {
        "operationId": "AccountController_create",
        "summary": "Create Account",
        "description": "This endpoint is used to create a new account with the provided details. You need to include the account details in the request body, and the API will return the created account object.\n\n**Request Body Parameters**:\n- **account_name** (required): The name of the account to be created.\n\n**How to Use**:\n- Send a POST request to the `/api/v1/admin/accounts` endpoint with the account details as a JSON object in the request body.\n- Ensure all required fields are included and correctly formatted.\n\n**Response Details**:\n- On successful creation, the API returns the newly created account object with its generated ID.\n- If any required field is missing or the provided data is invalid, the API responds with a 400 Bad Request error.\n- Server errors or permission issues will result in a 500 Internal Server Error.\n\nThis endpoint allows adding new accounts to the system, enabling organizational management.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ]
      },
      "get": {
        "operationId": "AccountController_findAll",
        "summary": "Get Page of Account List",
        "description": "This endpoint retrieves a paginated list of all accounts stored in the system. You can filter the results based on various parameters, allowing you to manage and view accounts effectively.\n\n**Query Parameters**:\n- **page** (optional): The page number of the results to retrieve (default is 1).\n- **limit** (optional): The number of accounts per page (default is 10).\n- **sort** (optional): The field by which to sort the accounts (e.g., account_name).\n- **order** (optional): The sorting order, either ascending or descending.\n\n**How to Use**:\n- Send a GET request to the `/api/v1/admin/accounts` endpoint with the desired query parameters.\n- Ensure parameters are correctly formatted for filtering and sorting.\n\n**Response Details**:\n- The API returns a JSON object containing a list of accounts along with pagination metadata (total count, current page, etc.).\n- If no accounts match the criteria, an empty list is returned.\n\nThis endpoint is useful for retrieving and displaying all accounts for better organizational management.",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter conditions",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/admin/accounts/{id}": {
      "get": {
        "operationId": "AccountController_findOne",
        "summary": "Get Account by account id",
        "description": "This endpoint retrieves the details of a specific account identified by its unique ID. It allows users to view account information in detail.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the account to retrieve.\n\n**How to Use**:\n- Send a GET request to the `/api/v1/admin/accounts/{id}` endpoint, replacing `{id}` with the actual account ID.\n- Ensure the ID is correctly formatted.\n\n**Response Details**:\n- On success, the API returns a JSON object with the account’s details.\n- If the account with the specified ID is not found, a 404 Not Found error is returned.\n\nThis endpoint is essential for accessing detailed information about specific accounts.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "operationId": "AccountController_update",
        "summary": "Update Account",
        "description": "This endpoint updates the details of an existing account identified by its unique ID. You need to provide the updated account details in the request body.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the account to update.\n\n**Request Body Parameters**:\n- **account_name** (optional): The updated name of the account.\n\n**How to Use**:\n- Send a PUT request to the `/api/v1/admin/accounts/{id}` endpoint with the updated account details as a JSON object in the request body.\n- Ensure that the account ID is in the URL and the request body includes valid fields.\n\n**Response Details**:\n- On successful update, the API returns the updated account object.\n- If the account is not found, a 404 Not Found error is returned.\n- Validation errors will result in a 400 Bad Request error.\n\nThis endpoint is crucial for maintaining accurate and up-to-date account information.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ]
      },
      "delete": {
        "operationId": "AccountController_delete",
        "summary": "Delete Account",
        "description": "This endpoint deletes an existing account from the system using its unique ID. It is a destructive operation that permanently removes the account.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the account to delete.\n\n**How to Use**:\n- Send a DELETE request to the `/api/v1/admin/accounts/{id}` endpoint, replacing `{id}` with the actual account ID.\n- Ensure that you have the necessary permissions to perform this operation.\n\n**Response Details**:\n- On successful deletion, the API returns a confirmation message indicating the account has been deleted.\n- If the account is not found, a 404 Not Found error is returned.\n- Permission errors will result in a 403 Forbidden error.\n\nThis endpoint is essential for managing account lifecycles by allowing deletions of obsolete or incorrect accounts.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/admin/permissions": {
      "post": {
        "operationId": "PermissionController_create",
        "summary": "Create Permission",
        "description": "This endpoint is used to create a new permission with the provided details. You need to include the permission details in the request body, and the API will return the created permission object.\n\n**Request Body Parameters**:\n- **permission_name** (required): The name of the permission (max length: 80 characters).\n- **permission_status** (optional): The status of the permission (default is \"active\").\n\n**How to Use**:\n- Send a POST request to the `/api/v1/admin/permissions` endpoint with the permission details as a JSON object in the request body.\n- Ensure all required fields are included and correctly formatted.\n\n**Response Details**:\n- On successful creation, the API returns the newly created permission object with its generated ID.\n- If any required field is missing or the provided data is invalid, the API responds with a 400 Bad Request error.\n- Server errors or permission issues will result in a 500 Internal Server Error.\n\nThis endpoint allows adding new permissions to the system.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "permissions"
        ]
      },
      "get": {
        "operationId": "PermissionController_findAll",
        "summary": "Retrieve All Permissions",
        "description": "This endpoint retrieves a list of all active permissions stored in the system. You can filter the results based on various parameters for better management.\n\n**How to Use**:\n- Send a GET request to the `/api/v1/admin/permissions` endpoint.\n- Ensure that any filtering parameters are correctly formatted, if applicable.\n\n**Response Details**:\n- The API returns a JSON object containing a list of permissions.\n- If no permissions are found, an empty list is returned.\n\nThis endpoint is useful for retrieving all permissions for better access control management.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "permissions"
        ]
      }
    },
    "/api/v1/admin/permissions/{id}": {
      "get": {
        "operationId": "PermissionController_findOne",
        "summary": "Retrieve Permission by ID",
        "description": "This endpoint retrieves the details of a specific permission identified by its unique ID. It allows users to view permission information in detail.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the permission to retrieve.\n\n**How to Use**:\n- Send a GET request to the `/api/v1/admin/permissions/{id}` endpoint, replacing `{id}` with the actual permission ID.\n- Ensure the ID is correctly formatted.\n\n**Response Details**:\n- On success, the API returns a JSON object with the permission’s details.\n- If the permission with the specified ID is not found, a 404 Not Found error is returned.\n\nThis endpoint is essential for accessing detailed information about specific permissions.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "permissions"
        ]
      },
      "patch": {
        "operationId": "PermissionController_update",
        "summary": "Update Permission",
        "description": "This endpoint updates the details of an existing permission identified by its unique ID. You need to provide the updated permission details in the request body.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the permission to update.\n\n**Request Body Parameters**:\n- **permission_name** (optional): The updated name of the permission (max length: 80 characters).\n- **permission_status** (optional): The updated status of the permission.\n\n**How to Use**:\n- Send a PUT request to the `/api/v1/admin/permissions/{id}` endpoint with the updated permission details as a JSON object in the request body.\n- Ensure that the permission ID is in the URL and the request body includes valid fields.\n\n**Response Details**:\n- On successful update, the API returns a confirmation message indicating the permission has been updated.\n- If the permission is not found, a 404 Not Found error is returned.\n- Validation errors will result in a 400 Bad Request error.\n\nThis endpoint is crucial for maintaining accurate and up-to-date permission information.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "permissions"
        ]
      },
      "delete": {
        "operationId": "PermissionController_delete",
        "summary": "Delete Permission",
        "description": "This endpoint deletes an existing permission from the system using its unique ID. It is a destructive operation that permanently removes the permission.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the permission to delete.\n\n**How to Use**:\n- Send a DELETE request to the `/api/v1/admin/permissions/{id}` endpoint, replacing `{id}` with the actual permission ID.\n- Ensure that you have the necessary permissions to perform this operation.\n\n**Response Details**:\n- On successful deletion, the API returns a confirmation message indicating the permission has been deleted.\n- If the permission is not found, a 404 Not Found error is returned.\n- Permission errors will result in a 403 Forbidden error.\n\nThis endpoint is essential for managing permission lifecycles by allowing deletions of obsolete or incorrect permissions.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "permissions"
        ]
      }
    },
    "/api/v1/admin/permissions/{id}/inactivate": {
      "patch": {
        "operationId": "PermissionController_inactivate",
        "summary": "Inactivate Permission",
        "description": "This endpoint inactivates an existing permission identified by its unique ID. It updates the status of the permission to inactive without deleting it.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the permission to inactivate.\n\n**How to Use**:\n- Send a PATCH request to the `/api/v1/admin/permissions/{id}/inactivate` endpoint, replacing `{id}` with the actual permission ID.\n- Ensure that you have the necessary permissions to perform this operation.\n\n**Response Details**:\n- On successful inactivation, the API returns a confirmation message indicating the permission has been inactivated.\n- If the permission is not found, a 404 Not Found error is returned.\n- Permission errors will result in a 403 Forbidden error.\n\nThis endpoint allows for the temporary disabling of permissions without deletion.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "permissions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Admin Service",
    "description": "Admin Service - admin services handle account, project, user and role management",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "schemas": {
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "account_name": {
            "type": "string",
            "description": "Name of account to be created"
          }
        },
        "required": [
          "account_name"
        ]
      },
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "account_name": {
            "type": "string",
            "description": "Name of account to be created"
          }
        }
      },
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "permission_name": {
            "type": "string",
            "description": "Name of permission to be created"
          }
        },
        "required": [
          "permission_name"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "permission_name": {
            "type": "string",
            "description": "Name of permission to be created"
          }
        }
      },
...
    }
  }
}